% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as-piechart-data.R
\name{as_piechart_data}
\alias{as_piechart_data}
\alias{as_piechart_data.matrix}
\alias{as_piechart_data.dist}
\title{Coerce to a Piechart data
Functions to coerce an object to a piechart data if possible.}
\usage{
as_piechart_data(x, ...)

\method{as_piechart_data}{matrix}(
  x,
  start = 0,
  end = 360,
  r0 = 0.5,
  r1 = 1,
  row_names = NULL,
  col_names = NULL,
  extra_mat = list(),
  ...
)

\method{as_piechart_data}{dist}(x, ...)
}
\arguments{
\item{x}{any \code{R} object.}

\item{...}{extra parameters.}

\item{start, end}{offset of starting and ending point from 3 o'clock in degree.}

\item{r0, r1}{start and end radius of arc heatmap.}

\item{row_names, col_names}{character vector.}

\item{extra_mat}{named list of extra matrix data.}
}
\value{
a piechart data.
}
\description{
Coerce to a Piechart data
Functions to coerce an object to a piechart data if possible.
}
\examples{
as_piechart_data(as.matrix(mtcars))
}
