% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom-arcbar.R
\name{geom_arcbar}
\alias{geom_arcbar}
\title{Arcbar Geom}
\usage{
geom_arcbar(
  mapping = NULL,
  data = NULL,
  inherit.aes = TRUE,
  width = 0.9,
  r0 = 0.5,
  r1 = 1,
  start = 60,
  end = 300,
  ID = NULL,
  ...
)
}
\arguments{
\item{mapping}{Set of aesthetic mappings.}

\item{data}{a data frame.}

\item{inherit.aes}{If FALSE, overrides the default aesthetics, rather than
combining with them.}

\item{width}{a numeric value between 0 to 1.}

\item{r0, r1}{a positive numeric value.}

\item{start, end}{a numeric value between 0 to 360.}

\item{ID}{character.}

\item{...}{other parameters.}
}
\description{
Geom function for draw arcbar.
}
\section{Aesthetics}{

\code{geom_arcbar()} understands the following aesthetics (required aesthetics are in bold):
    \itemize{
      \item \strong{\code{label}}
      \item \strong{\code{value}}
      \item \code{alpha}
      \item \code{fill}
      \item \code{colour}
      \item \code{size}
      \item \code{linetype}
   }
}

