% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom-arclink.R
\name{geom_arclink}
\alias{geom_arclink}
\title{Quadline Geom}
\usage{
geom_arclink(
  mapping = NULL,
  data = NULL,
  r = 0.98,
  from = NULL,
  to = NULL,
  ...
)
}
\arguments{
\item{mapping}{Set of aesthetic mappings.}

\item{data}{a data frame.}

\item{r}{a numeric value between 0 to 1.}

\item{from, to}{specifies the start and end ID of the connection.}

\item{...}{other parameters.}
}
\description{
Geom function for draw quadline.
}
\section{Aesthetics}{

\code{geom_arclink()} understands the following aesthetics (required aesthetics are in bold):
    \itemize{
      \item \strong{\code{from}}
      \item \strong{\code{to}}
      \item \code{alpha}
      \item \code{colour}
      \item \code{size}
      \item \code{linetype}
   }
}

