% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom-arcpoint.R
\name{geom_arcpoint}
\alias{geom_arcpoint}
\title{Arcpoint Geom}
\usage{
geom_arcpoint(
  mapping = NULL,
  data = NULL,
  inherit.aes = TRUE,
  jitter = NULL,
  r0 = 0.5,
  r1 = 1,
  start = 60,
  end = 300,
  ID = NULL,
  ...
)
}
\arguments{
\item{mapping}{Set of aesthetic mappings.}

\item{data}{a data frame.}

\item{inherit.aes}{If FALSE, overrides the default aesthetics, rather than
combining with them.}

\item{jitter}{one of "x", "y" or "xy".}

\item{r0, r1}{a positive numeric value.}

\item{start, end}{a numeric value between 0 to 360.}

\item{ID}{character.}

\item{...}{other parameters.}
}
\description{
Geom function for draw points on a arc canvas.
}
\section{Aesthetics}{

\code{geom_arcpoint()} understands the following aesthetics (required aesthetics are in bold):
    \itemize{
      \item \strong{\code{x}}
      \item \strong{\code{y}}
      \item \code{shape}
      \item \code{alpha}
      \item \code{fill}
      \item \code{colour}
      \item \code{size}
      \item \code{stroke}
   }
}

