% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom-pie2.R
\docType{data}
\name{geom_pie2}
\alias{geom_pie2}
\alias{GeomPie2}
\title{Pie2 Geom
Function to draw scatter piechart.}
\usage{
geom_pie2(
  mapping = NULL,
  data = NULL,
  stat = "identity",
  position = "identity",
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE,
  sort_by = NULL,
  decreasing = TRUE,
  start = 90,
  steps = 0.001,
  sum_value = NULL,
  label.facing = "downward",
  label.size = 3.88,
  label.col = "black",
  ...
)
}
\arguments{
\item{mapping}{Set of aesthetic mappings created by \code{\link[ggplot2:aes]{aes()}} or
\code{\link[ggplot2:aes_]{aes_()}}. If specified and \code{inherit.aes = TRUE} (the
default), it is combined with the default mapping at the top level of the
plot. You must supply \code{mapping} if there is no plot mapping.}

\item{data}{The data to be displayed in this layer. There are three
options:

If \code{NULL}, the default, the data is inherited from the plot
data as specified in the call to \code{\link[ggplot2:ggplot]{ggplot()}}.

A \code{data.frame}, or other object, will override the plot
data. All objects will be fortified to produce a data frame. See
\code{\link[ggplot2:fortify]{fortify()}} for which variables will be created.

A \code{function} will be called with a single argument,
the plot data. The return value must be a \code{data.frame}, and
will be used as the layer data. A \code{function} can be created
from a \code{formula} (e.g. \code{~ head(.x, 10)}).}

\item{stat}{The statistical transformation to use on the data for this
layer, as a string.}

\item{position}{Position adjustment, either as a string, or the result of
a call to a position adjustment function.}

\item{na.rm}{If \code{FALSE}, the default, missing values are removed with
a warning. If \code{TRUE}, missing values are silently removed.}

\item{show.legend}{logical. Should this layer be included in the legends?
\code{NA}, the default, includes if any aesthetics are mapped.
\code{FALSE} never includes, and \code{TRUE} always includes.
It can also be a named logical vector to finely select the aesthetics to
display.}

\item{inherit.aes}{If \code{FALSE}, overrides the default aesthetics,
rather than combining with them. This is most useful for helper functions
that define both data and aesthetics and shouldn't inherit behaviour from
the default plot specification, e.g. \code{\link[ggplot2:borders]{borders()}}.}

\item{sort_by}{a character vector of variable name.}

\item{decreasing}{logical. Should the sort order be increasing or decreasing?}

\item{start}{offset of starting point from 3 o'clock in radians.}

\item{steps}{increment of the sequence in radians.}

\item{sum_value}{sum of the value.}

\item{label.facing}{one of downward, binding or clockwise.}

\item{label.size}{the size of label.}

\item{label.col}{the color of label.}

\item{...}{Other arguments passed on to \code{\link[ggplot2:layer]{layer()}}. These are
often aesthetics, used to set an aesthetic to a fixed value, like
\code{colour = "red"} or \code{size = 3}. They may also be parameters
to the paired geom/stat.}
}
\description{
Pie2 Geom
Function to draw scatter piechart.
}
\section{Aesthetics}{

\code{geom_pie2()} understands the following aesthetics (required aesthetics are in bold):
    \itemize{
      \item \strong{\code{x}}
      \item \strong{\code{y}}
      \item \strong{\code{value}}
      \item \code{r0}
      \item \code{r1}
      \item \code{radius}
      \item \code{label}
      \item \code{alpha}
      \item \code{colour}
      \item \code{fill}
      \item \code{linetype}
      \item \code{size}
      \item \code{group}
   }
}

\keyword{datasets}
