% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/piechart-data-utils.R
\name{get_polygon_data}
\alias{get_polygon_data}
\alias{get_label_data}
\alias{calc_text_angle}
\alias{filter_generate}
\alias{pie_label}
\title{Helper Functions}
\usage{
get_polygon_data(x, ...)

get_label_data(x, ...)

calc_text_angle(.angle, facing = "clockwise", nice_facing = TRUE)

filter_generate(...)

pie_label(
  .label = NA,
  .ratio = NA,
  percent = TRUE,
  digits = 2,
  wrap = "(",
  ...
)
}
\arguments{
\item{x}{a piechart data.}

\item{...}{extra parameters passed on to \code{dplyr::filter()}.}

\item{.angle}{the angle in radians.}

\item{facing}{one of clockwise, binding or normal.}

\item{nice_facing}{logical value.}

\item{.label}{character vector.}

\item{.ratio}{numeric vector.}

\item{percent}{logical, if TRUE will convert ratio to percent.}

\item{digits}{integer indicating the number of decimal places.}

\item{wrap}{symbols.}
}
\value{
a data frame.
}
\description{
Helper functions for handle piechart data
}
\examples{
df <- piechart_data(mtcars, aes(value = wt))
get_polygon_data(df)
get_label_data(df)
}
