% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/piechart-package.R
\docType{package}
\name{piechart-package}
\alias{piechart-package}
\alias{_PACKAGE}
\title{piechart: Tools for Draw Piecharts}
\description{
The 'piechart' package can be used to draw a piecharts
    simply and directly based on 'ggplot2'.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/Hy4m/piechart}
  \item Report bugs at \url{https://github.com/Hy4m/piechart/issues}
}

}
