% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/piechart.R
\name{piechart}
\alias{piechart}
\title{Piechat Plot}
\usage{
piechart(
  data = NULL,
  mapping = aes(),
  xlim = NULL,
  ylim = NULL,
  clip = "on",
  ...
)
}
\arguments{
\item{data}{a data frame.}

\item{mapping}{default list of aesthetic mappings to use for plot.}

\item{xlim, ylim}{Limits for the x and y axes.}

\item{clip}{Should drawing be clipped to the extent of the plot panel?}

\item{...}{extra parameters passed on to \code{piechart_data()}.}
}
\value{
a ggplot object.
}
\description{
Init piechart plot
}
