% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/piechart-data.R
\name{piechart_data}
\alias{piechart_data}
\title{Piechat data}
\usage{
piechart_data(
  data,
  mapping,
  facet = NULL,
  sort_by = NULL,
  decreasing = TRUE,
  start = 90,
  steps = 0.01,
  sum_value = NULL,
  ...
)
}
\arguments{
\item{data}{a data frame.}

\item{mapping}{default list of aesthetic mappings to use for plot.}

\item{facet}{NULL or a formula.}

\item{sort_by}{a character vector of variable name.}

\item{decreasing}{logical. Should the sort order be increasing or decreasing?}

\item{start}{offset of starting point from 3 o'clock in degree.}

\item{steps}{increment of the sequence in radians.}

\item{sum_value}{sum of the value.}

\item{...}{extra parameters.}
}
\value{
a data frame.
}
\description{
Init piechart data
}
\examples{
library(ggplot2)
piechart_data(mtcars, aes(value = wt))
}
