% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pieGrob.R
\name{pieGrob}
\alias{pieGrob}
\alias{grid.pie}
\title{Draw Piechart}
\usage{
pieGrob(
  value = runif(10),
  label = NULL,
  x = 0.5,
  y = 0.5,
  r0 = 0.5,
  r1 = 1,
  start = 0,
  end = 360,
  sep = 0,
  steps = 0.001,
  radius = 10,
  label.facing = "downward",
  label.size = 12,
  label.col = "black",
  default.units = "native",
  gp = gpar()
)

grid.pie(...)
}
\arguments{
\item{value}{non-negtive numeric vector.}

\item{label}{label for each sector.}

\item{x, y}{a scala numeric value.}

\item{r0, r1}{the radius of piechart, should be between 0 and 1.}

\item{start, end}{offset of starting point from 3 o'clock in degrees.}

\item{sep}{gap bewteen two adjacent sectors.}

\item{steps}{increment of the sequence in radians.}

\item{radius}{the size (in mm) of pie.}

\item{label.facing}{one of downward, binding or clockwise.}

\item{label.size}{the size of label.}

\item{label.col}{the color of label.}

\item{default.units}{a string indicating the default units.}

\item{gp}{an object of class gpar.}

\item{...}{extra parameters.}
}
\value{
a polygonGrob.
}
\description{
These functions use to create and draw piechart.
}
