% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random-char.R
\name{random_char}
\alias{random_char}
\title{Random Words
Function to generate random words.}
\usage{
random_char(n, max_len = 10, min_len = 2, lambda = 5)
}
\arguments{
\item{n}{number of random words to return.}

\item{max_len, min_len}{maximum and minimum length of random words.}

\item{lambda}{positive integer value.}
}
\value{
character vector.
}
\description{
Random Words
Function to generate random words.
}
\examples{
random_char(10)
random_char(10, min_len = 3)
}
