% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default-theme.R
\name{theme_piechart}
\alias{theme_piechart}
\title{Default Theme}
\usage{
theme_piechart(...)
}
\arguments{
\item{...}{extra parameters passed on to \code{theme()}.}
}
\value{
a complete theme.
}
\description{
Default theme for piechart
}
