% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zoom-piechart.R
\name{zoom_piechart}
\alias{zoom_piechart}
\title{Zoom Piechart data
Zoom the specific area according to the conditions.}
\usage{
zoom_piechart(data, ..., zoom = 2, base_line = NULL)
}
\arguments{
\item{data}{a piechart data.}

\item{...}{any conditions.}

\item{zoom}{numeric, should be greater than 1.}

\item{base_line}{numeric, should be greater than r0 and less than r1 * zoom.}
}
\value{
a piechart data.
}
\description{
Zoom Piechart data
Zoom the specific area according to the conditions.
}
