% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation.R
\name{anno_rho_axis}
\alias{anno_rho_axis}
\alias{anno_alpha_axis}
\alias{anno_rect}
\title{Annotation for Piechart
Functions to draw axis, rect and so on.}
\usage{
anno_rho_axis(
  breaks,
  labels,
  scale,
  r = 0.5,
  start = 0,
  end = 360,
  dtick = "inside",
  ltick = 0.01,
  lcol = "black",
  lsize = 0.5,
  dtext = "inside",
  tcol = "black",
  tsize = 3.5,
  facing = "binding",
  nice_facing = TRUE,
  gap = 0.02,
  ...
)

anno_alpha_axis(
  breaks,
  labels,
  scale,
  r0 = 0.5,
  r1 = 1,
  angle = 90,
  dtick = "left",
  ltick = 1,
  lcol = "black",
  lsize = 0.5,
  dtext = "left",
  tcol = "black",
  tsize = 3.5,
  gap = 1,
  ...
)

anno_rect(r0 = 0.5, r1 = 1, start = 90, end = 270, ...)
}
\arguments{
\item{breaks}{the points at which tick-marks are to be drawn.}

\item{labels}{a character or expression vector of labels to be
placed at the tickpoints.}

\item{scale}{origin range of data.}

\item{r, r0, r1}{positive value.}

\item{dtick, dtext}{direction of ticks and labels.}

\item{ltick}{length of ticks.}

\item{lcol, lsize}{axis's colour and line width.}

\item{tcol, tsize}{label's colour and size.}

\item{facing}{one of clockwise, binding or normal.}

\item{nice_facing}{logical value.}

\item{gap}{gap between ticks and labels.}

\item{...}{extra parameters.}

\item{angle, start, end}{the angle in radians.}
}
\value{
a gg object.
}
\description{
Annotation for Piechart
Functions to draw axis, rect and so on.
}
